[gh-41]: refactor actions
