+++ src/app/core/actions/action-creators/stored-skipasses.action-creators.js
@@ -1,9 +1,4 @@
 import {
-  save as saveSkipass,
-  remove as removeSkipass,
-  getAll as getStoredSkipasses
-} from '../data-access/skipassRepository'
-import {
   STORE_SKIPASS_REQUEST,
   STORE_SKIPASS_SUCCESS,
   STORE_SKIPASS_FAILURE,
@@ -13,7 +8,7 @@ import {
   REMOVE_STORED_SKIPASS_REQUEST,
   REMOVE_STORED_SKIPASS_SUCCESS,
   REMOVE_STORED_SKIPASS_FAILURE
-} from './actionTypes'
+} from '../action-types'
 
 const storeSkipassRequest = () => {
   return {
@@ -74,49 +69,14 @@ const removeStoredSkipassFailure = error => {
   }
 }
 
-const storeSkipass = () => {
-  return async (dispatch, getState) => {
-    const skipass = getState().get('searchSkipass').get('skipass').toJS()
-
-    dispatch(storeSkipassRequest())
-
-    try {
-      await saveSkipass(skipass)
-      dispatch(storeSkipassSuccess())
-    } catch (e) {
-      dispatch(storeSkipassFailure(e))
-    }
-  }
-}
-
-const fetchStoredSkipasses = () => {
-  return async (dispatch) => {
-    dispatch(fetchStoredSkipassesRequest())
-
-    try {
-      const skipasses = await getStoredSkipasses()
-      dispatch(fetchStoredSkipassesSuccess(skipasses))
-    } catch (e) {
-      dispatch(fetchStoredSkipassesFailure(e))
-    }
-  }
-}
-
-const removeStoredSkipass = skipass => {
-  return async dispatch => {
-    dispatch(removeStoredSkipassRequest())
-
-    try {
-      await removeSkipass(skipass)
-      dispatch(removeStoredSkipassSuccess(skipass))
-    } catch (e) {
-      dispatch(removeStoredSkipassFailure(e))
-    }
-  }
-}
-
 export {
-  storeSkipass,
-  fetchStoredSkipasses,
-  removeStoredSkipass
+  storeSkipassRequest,
+  storeSkipassSuccess,
+  storeSkipassFailure,
+  fetchStoredSkipassesRequest,
+  fetchStoredSkipassesSuccess,
+  fetchStoredSkipassesFailure,
+  removeStoredSkipassRequest,
+  removeStoredSkipassSuccess,
+  removeStoredSkipassFailure
 }
